added screen coordinate calculation in video processing and restructured a bit
