+++ .gitignore
@@ -1,6 +1,5 @@
 .DS_Store
 models
-output
 no-commit
 
 # Byte-compiled / optimized / DLL files
