+++ processing/__init__.py
@@ -1,0 +1,3 @@
+from .image import process_image
+from .video import process_video
+from .webcam import process_webcam
